requirements.txt
asgiref==3.5.0
Django==4.0.4
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.47
Pillow==9.1.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
